transormation du fichier Slideshow en tsx + ajout du typage
